requirements.txt
Flask==2.2.5
numpy==1.25.2
pandas==2.0.3
pandas-datareader==0.10.0
pandas-gbq==0.19.2
pandas-stubs==2.0.3.230814
joblib==1.4.2
xgboost==2.0.3
sklearn-pandas==2.2.0
scipy==1.11.4
shap==0.46.0